+++ app/src/main/java/com/sedilant/yambol/domain/get/GetLastTrainOfTeamUseCase.kt
@@ -1,5 +1,5 @@
 package com.sedilant.yambol.domain.get
 
 interface GetLastTrainOfTeamUseCase {
-    suspend operator fun invoke(teamId: Int?): Long?
+    suspend operator fun invoke(teamId: Long?): Long?
 }
