+++ app/src/main/java/com/sedilant/yambol/domain/get/GetLastTrainOfTeamUseCaseImpl.kt
@@ -6,7 +6,7 @@ import javax.inject.Inject
 class GetLastTrainOfTeamUseCaseImpl @Inject constructor(
     private val teamRepository: TeamRepository
 ) : GetLastTrainOfTeamUseCase {
-    override suspend fun invoke(teamId: Int?): Long? {
+    override suspend fun invoke(teamId: Long?): Long? {
         return if (teamId != null) {
             teamRepository.getLastTrainWithTrainTaskByTeamId(teamId)
         } else null
