+++ app/src/main/java/com/sedilant/yambol/domain/get/GetTeamObjectivesUseCase.kt
@@ -4,5 +4,5 @@ import com.sedilant.yambol.domain.models.TeamObjectivesDomainModel
 import kotlinx.coroutines.flow.Flow
 
 interface GetTeamObjectivesUseCase {
-    suspend operator fun invoke(teamId: Int): Flow<List<TeamObjectivesDomainModel>>
+    suspend operator fun invoke(teamId: Long): Flow<List<TeamObjectivesDomainModel>>
 }
