+++ app/src/main/java/com/sedilant/yambol/domain/models/TrainDomainModel.kt
@@ -7,5 +7,5 @@ data class TrainDomainModel(
     val date: Date,
     val time: Float,
     val concepts: List<String>,
-    val teamId: Int
+    val teamId: Long
 )
