+++ app/src/main/java/com/sedilant/yambol/domain/models/TrainTaskDomainModel.kt
@@ -1,7 +1,7 @@
 package com.sedilant.yambol.domain.models
 
 data class TrainTaskDomainModel(
-    val trainingTaskId: Int,
+    val trainingTaskId: Long,
     val name: String,
     val concept: String,
     val description: String,
